+++ src/lib/Block/Block.svelte
@@ -19,12 +19,20 @@
 <style>
   .layout-block {
     margin: 0 auto;
-    width: 100%;
   }
   .layout-block.body {
     max-width: var(--body-max-width, 760px);
+    margin: 0 var(--spacing-4);
   }
   .layout-block.full {
     max-width: none;
+    width: 100%;
+  }
+
+  @media screen and (min-width: 792px) {
+    .layout-block.body {
+      padding: 0;
+      margin: 0 auto;
+    }
   }
 </style>